+++ src/Web/HttpRequest.cpp
@@ -12,53 +12,49 @@ HttpRequest::HttpRequest(QObject* parent)
 {
 }
 
+QNetworkAccessManager* HttpRequest::manager()
+{
+    if (this->networkManager == nullptr)
+        this->networkManager = new QNetworkAccessManager(this);
+
+    return this->networkManager;
+}
+
+// Each reply carries its own URL to the log, so two requests in flight at once
+// are each logged as themselves.
+void HttpRequest::logReply(const QString& method, const QString& url, QNetworkReply* reply)
+{
+    QString data = QString::fromUtf8(reply->readAll());
+    int statusCode = reply->attribute(QNetworkRequest::HttpStatusCodeAttribute).toInt();
+
+    qDebug("HttpRequest::send%sFinished %d %s", method == "GET" ? "Get" : "Post", statusCode, qPrintable(data));
+
+    HttpResponseLog::getInstance().logResponse(method, url, statusCode, data);
+
+    reply->deleteLater();
+}
+
 void HttpRequest::sendGet(const QString& url, const QUrlQuery& query)
 {
     QUrl request(url);
     request.setQuery(query);
 
-    this->pendingGetUrl = request.toString();
+    const QString sentUrl = request.toString();
 
-    qDebug("HttpRequest::sendGet %s", qPrintable(this->pendingGetUrl));
+    qDebug("HttpRequest::sendGet %s", qPrintable(sentUrl));
 
-    this->networkManager = new QNetworkAccessManager(this);
-    QObject::connect(this->networkManager, SIGNAL(finished(QNetworkReply*)), this, SLOT(sendGetFinished(QNetworkReply*)));
-    this->networkManager->get(QNetworkRequest(request));
-}
-
-void HttpRequest::sendGetFinished(QNetworkReply* reply)
-{
-    QString data = QString::fromUtf8(reply->readAll());
-    int statusCode = reply->attribute(QNetworkRequest::HttpStatusCodeAttribute).toInt();
-
-    qDebug("HttpRequest::sendGetFinished %d %s", statusCode, qPrintable(data));
-
-    HttpResponseLog::getInstance().logResponse("GET", this->pendingGetUrl, statusCode, data);
-
-    reply->deleteLater();
-    this->networkManager->deleteLater();
+    QNetworkReply* reply = manager()->get(QNetworkRequest(request));
+    QObject::connect(reply, &QNetworkReply::finished, this, [this, reply, sentUrl]() {
+        logReply("GET", sentUrl, reply);
+    });
 }
 
 void HttpRequest::sendPost(const QString& url, const QUrlQuery& query)
 {
-    this->pendingPostUrl = url;
-
     qDebug("HttpRequest::sendPost %s, %s", qPrintable(url), qPrintable(query.toString(QUrl::FullyEncoded)));
 
-    this->networkManager = new QNetworkAccessManager(this);
-    QObject::connect(this->networkManager, SIGNAL(finished(QNetworkReply*)), this, SLOT(sendPostFinished(QNetworkReply*)));
-    this->networkManager->post(QNetworkRequest(QUrl(url)), query.toString(QUrl::FullyEncoded).toUtf8());
-}
-
-void HttpRequest::sendPostFinished(QNetworkReply* reply)
-{
-    QString data = QString::fromUtf8(reply->readAll());
-    int statusCode = reply->attribute(QNetworkRequest::HttpStatusCodeAttribute).toInt();
-
-    qDebug("HttpRequest::sendPostFinished %d %s", statusCode, qPrintable(data));
-
-    HttpResponseLog::getInstance().logResponse("POST", this->pendingPostUrl, statusCode, data);
-
-    reply->deleteLater();
-    this->networkManager->deleteLater();
+    QNetworkReply* reply = manager()->post(QNetworkRequest(QUrl(url)), query.toString(QUrl::FullyEncoded).toUtf8());
+    QObject::connect(reply, &QNetworkReply::finished, this, [this, reply, url]() {
+        logReply("POST", url, reply);
+    });
 }